moved the ttl_delete test to a separate class
