+++ ydb/tests/olap/ttl_tiering/ya.make
@@ -3,12 +3,13 @@ ENV(YDB_DRIVER_BINARY="ydb/apps/ydbd/ydbd")
 ENV(MOTO_SERVER_PATH="contrib/python/moto/bin/moto_server")
 ENV(YDB_ADDITIONAL_LOG_CONFIGS="TX_TIERING:DEBUG")
 
-FORK_TEST_FILES()
+FORK_TESTS()
 
 TEST_SRCS(
     base.py
     data_correctness.py
     ttl_delete_s3.py
+    ttl_delete.py
     ttl_unavailable_s3.py
     data_migration_when_alter_ttl.py
     unstable_connection.py
